+++ src/main/java/org/folio/qm/service/impl/LinksServiceImpl.java
@@ -1,10 +1,12 @@
 package org.folio.qm.service.impl;
 
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.folio.qm.client.LinksClient;
 import org.folio.qm.client.LinksClient.InstanceLink;
 import org.folio.qm.client.LinksClient.InstanceLinks;
 import org.folio.qm.domain.dto.FieldItem;
+import org.folio.qm.domain.dto.LinkDetails;
 import org.folio.qm.domain.dto.MarcFormat;
 import org.folio.qm.domain.dto.QuickMarc;
 import org.folio.qm.service.LinkingRulesService;
@@ -44,12 +46,12 @@ public class LinksServiceImpl implements LinksService {
 
   private InstanceLinks extractLinks(QuickMarc quickMarc) {
     var links = quickMarc.getFields().stream()
-      .filter(fieldItem -> fieldItem.getAuthorityId() != null)
+      .filter(fieldItem -> fieldItem.getLinkDetails() != null)
       .map(fieldItem -> new InstanceLink()
         .setInstanceId(quickMarc.getExternalId())
-        .setAuthorityId(fieldItem.getAuthorityId())
-        .setAuthorityNaturalId(fieldItem.getAuthorityNaturalId())
-        .setLinkingRuleId(fieldItem.getLinkingRuleId()))
+        .setAuthorityId(fieldItem.getLinkDetails().getAuthorityId())
+        .setAuthorityNaturalId(fieldItem.getLinkDetails().getAuthorityNaturalId())
+        .setLinkingRuleId(fieldItem.getLinkDetails().getLinkingRuleId()))
       .toList();
 
     return new InstanceLinks(links, links.size());
@@ -68,7 +70,10 @@ public class LinksServiceImpl implements LinksService {
             populateLink(fields.get(0), instanceLink);
           } else {
             fields.stream()
-              .filter(fieldItem -> instanceLink.getAuthorityId().equals(fieldItem.getAuthorityId()))
+              .filter(fieldItem -> instanceLink.getAuthorityId().equals(
+                Optional.ofNullable(fieldItem.getLinkDetails())
+                  .map(LinkDetails::getAuthorityId)
+                  .orElse(null)))
               .forEach(fieldItem -> populateLink(fieldItem, instanceLink));
           }
         })
@@ -76,8 +81,11 @@ public class LinksServiceImpl implements LinksService {
   }
 
   private void populateLink(FieldItem fieldItem, InstanceLink instanceLink) {
-    fieldItem.setAuthorityId(instanceLink.getAuthorityId());
-    fieldItem.setAuthorityNaturalId(instanceLink.getAuthorityNaturalId());
-    fieldItem.setLinkingRuleId(instanceLink.getLinkingRuleId());
+    var linkDetails = fieldItem.getLinkDetails();
+    linkDetails.setAuthorityId(instanceLink.getAuthorityId());
+    linkDetails.setAuthorityNaturalId(instanceLink.getAuthorityNaturalId());
+    linkDetails.setLinkingRuleId(instanceLink.getLinkingRuleId());
+    linkDetails.setStatus(instanceLink.getStatus());
+    linkDetails.setErrorCause(instanceLink.getErrorCause());
   }
 }