+++ src/test/java/org/folio/qm/controller/RecordsEditorIT.java
@@ -40,6 +40,8 @@ import static org.folio.qm.support.utils.testentities.TestEntitiesUtils.EXISTED_
 import static org.folio.qm.support.utils.testentities.TestEntitiesUtils.JOB_EXECUTION_CREATED;
 import static org.folio.qm.support.utils.testentities.TestEntitiesUtils.JOHN_USER_ID;
 import static org.folio.qm.support.utils.testentities.TestEntitiesUtils.LINKING_RULE_ID;
+import static org.folio.qm.support.utils.testentities.TestEntitiesUtils.LINK_ERROR_CAUSE;
+import static org.folio.qm.support.utils.testentities.TestEntitiesUtils.LINK_STATUS_ERROR;
 import static org.folio.qm.support.utils.testentities.TestEntitiesUtils.PARSED_RECORD_AUTHORITY_DTO_PATH;
 import static org.folio.qm.support.utils.testentities.TestEntitiesUtils.PARSED_RECORD_BIB_DTO_PATH;
 import static org.folio.qm.support.utils.testentities.TestEntitiesUtils.PARSED_RECORD_HOLDINGS_DTO_PATH;
@@ -105,9 +107,11 @@ class RecordsEditorIT extends BaseIT {
       .andExpect(jsonPath("$.suppressDiscovery").value(Boolean.FALSE))
       .andExpect(jsonPath("$.parsedRecordId").value(VALID_PARSED_RECORD_ID.toString()))
       .andExpect(jsonPath("$.updateInfo.updatedBy.userId").value(JOHN_USER_ID))
-      .andExpect(jsonPath("$.fields[14].authorityId").value(AUTHORITY_ID))
-      .andExpect(jsonPath("$.fields[14].authorityNaturalId").value(AUTHORITY_NATURAL_ID))
-      .andExpect(jsonPath("$.fields[14].linkingRuleId").value(LINKING_RULE_ID));
+      .andExpect(jsonPath("$.fields[14].linkDetails.authorityId").value(AUTHORITY_ID))
+      .andExpect(jsonPath("$.fields[14].linkDetails.authorityNaturalId").value(AUTHORITY_NATURAL_ID))
+      .andExpect(jsonPath("$.fields[14].linkDetails.linkingRuleId").value(LINKING_RULE_ID))
+      .andExpect(jsonPath("$.fields[14].linkDetails.status").value(LINK_STATUS_ERROR))
+      .andExpect(jsonPath("$.fields[14].linkDetails.errorCause").value(LINK_ERROR_CAUSE));
 
     checkParseRecordDtoId();
   }